UT_ASSERTs moved to test/c_api
