+++ test/c_api/disjoint_pool.c
@@ -7,6 +7,7 @@
 #include "pool_disjoint.h"
 #include "provider_null.h"
 #include "test_helpers.h"
+#include "test_ut_asserts.h"
 
 void test_disjoint_pool_default_params(void) {
     umf_memory_provider_handle_t provider = nullProviderCreate();
